+++ database/seeders/PreLoadSeeder.php
@@ -23,15 +23,28 @@ class PreLoadSeeder extends Seeder
     {
 
         $sampleid = (string) Str::uuid();
+        $sampleuserid = (string) Str::uuid();
+
 
         DB::table('custom_user_tables')->insert([
-            'user_id' => $sampleid,
-            'email' => 'masteradmin@sample.edu.ph',
-            'password' => bcrypt('admin@123'),
-            'is_admin' => true,
-            'login_attempts' => 0,
-            'created_at' => Carbon::now('Asia/Manila'),
-            'updated_at' => Carbon::now('Asia/Manila')
+            [
+                'user_id' => $sampleid,
+                'email' => 'masteradmin@sample.edu.ph',
+                'password' => bcrypt('admin@123'),
+                'is_admin' => true,
+                'login_attempts' => 0,
+                'created_at' => Carbon::now('Asia/Manila'),
+                'updated_at' => Carbon::now('Asia/Manila')
+            ],
+            [
+                'user_id' => $sampleuserid,
+                'email' => 'user123@gmail.com',
+                'password' => bcrypt('user@123'),
+                'is_admin' => false,
+                'login_attempts' => 0,
+                'created_at' => Carbon::now('Asia/Manila'),
+                'updated_at' => Carbon::now('Asia/Manila')
+            ]
         ]);
 
         $officeid = (string) Str::uuid();
@@ -54,6 +67,17 @@ class PreLoadSeeder extends Seeder
             'office_id' => $officeid,
             'is_master_admin' => true,
             'is_technician' => true,
+            'created_at' => Carbon::now('Asia/Manila'),
+            'updated_at' => Carbon::now('Asia/Manila')
+        ]);
+
+        DB::table('user_profiles')->insert([
+            'profile_id' => (string) Str::uuid(),
+            'user_id' => $sampleuserid,
+            'last_name' => 'Cruz',
+            'first_name' => 'Juan',
+            'middle_name' => 'Dela',
+            'gender' => 'Male',
             'created_at' => Carbon::now('Asia/Manila'),
             'updated_at' => Carbon::now('Asia/Manila')
         ]);